+++ db/migrate/20210117143211_remove_services.rb
@@ -1,0 +1,6 @@
+class RemoveServices < ActiveRecord::Migration[6.0]
+  def change
+    remove_column :agents, :service_id
+    drop_table :services
+  end
+end