+++ spec/lib/agents_exporter_spec.rb
@@ -9,7 +9,7 @@ describe AgentsExporter do
     let(:tag_bg_color) { '#000000' }
     let(:icon) { 'Camera' }
     let(:agent_list) { [agents(:jane_status_agent), agents(:jane_notifier_agent)] }
-    let(:exporter) { 
+    let(:exporter) {
       AgentsExporter.new(
         agents: agent_list, name: name, description: description,
         guid: guid, tag_fg_color: tag_fg_color,
@@ -29,7 +29,7 @@ describe AgentsExporter do
       expect(Time.parse(data[:exported_at])).to be_within(2).of(Time.now.utc)
       expect(data[:links]).to eq([{ source: guid_order(agent_list, :jane_status_agent), receiver: guid_order(agent_list, :jane_notifier_agent) }])
       expect(data[:control_links]).to eq([])
-      expect(data[:agents]).to eq(agent_list.sort_by { |a| a.guid }.map { |agent| exporter.agent_as_json(agent) })
+      expect(data[:agents]).to eq(agent_list.sort_by(&:guid).map { |agent| exporter.agent_as_json(agent) })
       expect(data[:agents].all? { |agent_json| agent_json[:guid].present? && agent_json[:type].present? && agent_json[:name].present? }).to be_truthy
 
       expect(data[:agents][guid_order(agent_list, :jane_notifier_agent)]).not_to have_key(:schedule) # can't be scheduled
@@ -43,7 +43,7 @@ describe AgentsExporter do
     end
 
     it 'outputs control links to agents within the incoming set, but not outside it' do
-      agents(:jane_notifier_agent).control_targets = [agents(:jane_status_agent), agents(:jane_basecamp_agent)]
+      agents(:jane_notifier_agent).control_targets = [agents(:jane_status_agent), agents(:jane_website_agent)]
       agents(:jane_notifier_agent).save!
 
       expect(exporter.as_json[:control_links]).to eq([{ controller: guid_order(agent_list, :jane_notifier_agent), control_target: guid_order(agent_list, :jane_status_agent) }])
@@ -76,6 +76,6 @@ describe AgentsExporter do
   end
 
   def guid_order(agent_list, agent_name)
-    agent_list.map { |a| a.guid }.sort.find_index(agents(agent_name).guid)
+    agent_list.map(&:guid).sort.find_index(agents(agent_name).guid)
   end
 end