+++ spec/models/agents/http_status_agent_spec.rb
@@ -17,7 +17,6 @@ describe 'HttpStatusAgent' do
   let(:agent) do
     Agents::HttpStatusAgent.create!(
       name: SecureRandom.uuid,
-      service: services(:generic),
       user: users(:jane),
       options: agent_options
     )