+++ spec/support/shared_examples/agent_controller_concern.rb
@@ -131,11 +131,11 @@ shared_examples_for AgentControllerConcern do
 
     it 'should configure targets with nested objects' do
       agent.control_targets = [
-        agents(:bob_basecamp_agent), # does not support a `template` option, but anyway
+        agents(:bob_status_agent), # does not support a `template` option, but anyway
         agents(:bob_data_output_agent)
       ]
       agent.options['action'] = 'configure'
-      agent.options['configure_options'] = { 
+      agent.options['configure_options'] = {
         template: {
           item: {
             title: 'changed'